Added dialog confirmation for delete

+++ src/app/main-library/main-library.component.ts
@@ -1,12 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {DataSource} from '@angular/cdk/collections';
-import {Observable, ReplaySubject} from 'rxjs';
+import {Observable, ReplaySubject, map} from 'rxjs';
 import {MatTableModule} from '@angular/material/table';
 import {MatButtonModule} from '@angular/material/button';
-
+import { MatDialog } from '@angular/material/dialog';
 
 import { MainLibraryBookDataService } from '../main-library-data-service.service';
+import { DeleteConfirmationDialogComponent } from '../delete-confirmation-dialog/delete-confirmation-dialog.component';
 
 export interface Book {
   bookTitle: string;
@@ -26,12 +27,30 @@ export interface Book {
 
 })
 export class MainLibraryComponent implements OnInit {
-  constructor(private bookDataService: MainLibraryBookDataService) {}
+  constructor(private bookDataService: MainLibraryBookDataService, private dialog: MatDialog) {}
  
   displayedColumns: string[] = ['bookTitle', 'author', 'numberOfPages', 'genres', 'review', 'description'];
   dataToDisplay: Book[] = [];
   dataSource = new ExampleDataSource(this.dataToDisplay);
   selectedRowIndex: number = -1;
+
+  // openConfirmationDialog(): void {
+  //   this.bookDataService.getDataAtIndex(this.selectedRowIndex).subscribe((book: Book | null) => {
+  //     if (book !== null) {
+  //       const dialogRef = this.dialog.open(DeleteConfirmationDialogComponent, {
+  //         data: { bookTitle: book.bookTitle },
+  //       });
+  
+  //       dialogRef.afterClosed().subscribe((result) => {
+  //         if (result) {
+  //           this.removeData();
+  //         }
+  //       });
+  //     }
+  //   });
+  // }
+  
+
 
   ngOnInit(): void {
     console.log("Hello :)")
@@ -43,11 +62,27 @@ export class MainLibraryComponent implements OnInit {
     
   }
   
-  removeData() {
-    if (this.selectedRowIndex !== -1) {
-      this.bookDataService.removeEntry(this.selectedRowIndex);
-      this.selectedRowIndex = -1;
-    }
+  removeData(): void {
+    this.getBookTitle().subscribe((bookTitle: string | null) => {
+      if (this.selectedRowIndex !== -1 && bookTitle !== null) {
+        const dialogRef = this.dialog.open(DeleteConfirmationDialogComponent, {
+          data: { bookTitle: bookTitle },
+        });
+  
+        dialogRef.afterClosed().subscribe((result) => {
+          if (result) {
+            this.bookDataService.removeEntry(this.selectedRowIndex);
+            this.selectedRowIndex = -1;
+          }
+        });
+      }
+    });
+  }
+  
+  getBookTitle(): Observable<string | null> {
+    return this.bookDataService.getDataAtIndex(this.selectedRowIndex).pipe(
+      map((book: Book | null) => (book !== null ? book.bookTitle : null))
+    );
   }
   addData() {
